+++ src/Dapr.PluggableComponents/Components/SecretStore/SecretStoreBulkGetRequest.cs
@@ -13,13 +13,12 @@
 
 using Dapr.Proto.Components.V1;
 
-namespace Dapr.PluggableComponents.Components.SecretStores;
+namespace Dapr.PluggableComponents.Components.SecretStore;
 
 
 /// <summary>
 /// Represents properties associated with a request to retrieve all secrets from a secret store.
 /// </summary>
-/// <param name="secretName">The secret that should be retrieved.</param>
 public sealed record SecretStoreBulkGetRequest
 {
     /// <summary>
