+++ src/Dapr.PluggableComponents/Components/SecretStore/SecretStoreGetResponse.cs
@@ -14,12 +14,13 @@
 using Dapr.PluggableComponents.Utilities;
 using Dapr.Proto.Components.V1;
 using Google.Protobuf;
-namespace Dapr.PluggableComponents.Components.SecretStores;
+
+namespace Dapr.PluggableComponents.Components.SecretStore;
 
 /// <summary>
-/// Represents properties associated with a response to retrieving bulk secret from a secret store.
+/// Represents properties associated with a response to retrieving secret from a secret store.
 /// </summary>
-public sealed record SecretStoreResponse
+public sealed record SecretStoreGetResponse
 {
     /// <summary>
     /// Gets or sets the key's value.
@@ -30,18 +31,17 @@ public sealed record SecretStoreResponse
     public IReadOnlyDictionary<string, string> Data { get; init; } = new Dictionary<string, string>();
 
 
-    internal static SecretResponse ToGetResponse(SecretStoreResponse response)
+    internal static GetSecretResponse ToGetResponse(SecretStoreGetResponse response)
     {
-        var grpcResponse = new SecretResponse();
+        var grpcResponse = new GetSecretResponse();
 
         // NOTE: in case of not found, you should not return any error.
 
         if (response != null)
         {
-            grpcResponse.Secrets.Add(response.Data);
+            grpcResponse.Data.Add(response.Data);
         }
 
         return grpcResponse;
     }
 }
-
